agent/pyproject.toml
[project]
name = "server-provision-agent"
version = "0.1.0"
description = "LangGraph agent for the Smart Server Provisioning Assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pymongo>=4.7",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "langgraph>=0.2",
    "langgraph-checkpoint-mongodb>=0.1",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.2",
    "langchain-openai>=0.2",
    "copilotkit>=0.1.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
